pyproject.toml
[tool.poetry]
name = "simple-backups"
version = "1.4.0"
description = "A simple yet expandable backup script"
authors = ["Roselyn Coales <roselyn@coales.co.uk>"]

[tool.poetry.dependencies]
python = "^3.10"
schedule = "^1.1.0"
google-cloud-storage = "^1.39.0"
simple-heartbeat = "^0.1.2"
paramiko = "^2.7.2"
requests = "2.32.4"
prometheus-client = "^0.11.0"
click = "^8.2.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
